dns: divide header in three parts
