+++ src/dns/include.am
@@ -7,4 +7,6 @@ libmeasurement_kit_la_LIBADD += src/dns/libdns.la
 
 dnsincludedir = $(includedir)/measurement_kit/dns
 dnsinclude_HEADERS = \
-    include/measurement_kit/dns/dns.hpp
+    include/measurement_kit/dns/query.hpp \
+    include/measurement_kit/dns/response.hpp \
+    include/measurement_kit/dns/resolver.hpp
